backend/requirements.txt
# 🌍 GEOWISE - Geospatial AI Platform Requirements
# Core Framework & Async
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database & ORM (SQLite is built-in)
sqlalchemy==2.0.23
alembic==1.12.1

# Logging
structlog==23.2.0
loguru==0.7.2

# Environment & Configuration
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# HTTP Client & External APIs
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Geospatial Libraries
h3==3.7.7
shapely==2.0.2
geojson==3.1.0
pyproj==3.6.1

# Data Processing & Analysis
pandas==2.1.3
numpy==1.25.2
scipy==1.11.3
scikit-learn==1.3.2

# Raster Processing
rasterio==1.3.8
pillow==10.1.0

# Background Tasks & Cache
celery==5.3.4
redis==5.0.1

# Vector Database for RAG
chromadb==0.4.22
sentence-transformers==2.2.2

# LLM Integration
groq==0.4.0
openai==1.3.9
tiktoken==0.5.2

# Testing & Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
country-bounding-boxes==1.0.3
groq sentence-transformers 
chromadb
planetary-computer==1.0.0
pyproj>=3.0.0
earthengine-api --break-system-packages

google-auth --break-system-packages